Move common validation methods to base class (#37151)
