+++ src/python_testing/matter_testing_infrastructure/chip/testing/matter_testing.py
@@ -972,6 +972,46 @@ class MatterBaseTest(base_test.BaseTestClass):
         # The named pipe name must be set in the derived classes
         self.app_pipe = None
 
+    @staticmethod
+    def is_valid_uint_value(value, bit_count=64):
+        """
+        Checks if 'value' is a non-negative integer fitting into 'bit_count' bits.
+        For example, bit_count=32 => must fit within 0 <= value <= 0xFFFFFFFF
+        """
+        if not isinstance(value, int):
+            return False
+        if value < 0:
+            return False
+        return value < 2**bit_count
+
+    @staticmethod
+    def is_valid_str_value(value):
+        return isinstance(value, str) and len(value) > 0
+
+    def assert_valid_uint64(self, value, field_name):
+        """Asserts that the value is a valid uint64."""
+        asserts.assert_true(self.is_valid_uint_value(value, bit_count=64),
+                            f"{field_name} should be a uint64 or NULL.")
+
+    def assert_valid_uint32(self, value, field_name):
+        """Asserts that the value is a valid uint32."""
+        asserts.assert_true(self.is_valid_uint_value(value, bit_count=32),
+                            f"{field_name} should be a uint32 or NULL.")
+
+    def assert_valid_uint16(self, value, field_name):
+        """Asserts that the value is a valid uint16."""
+        asserts.assert_true(self.is_valid_uint_value(value, bit_count=16),
+                            f"{field_name} should be a uint16 or NULL.")
+
+    def assert_valid_uint8(self, value, field_name):
+        """Asserts that the value is a valid uint16."""
+        asserts.assert_true(self.is_valid_uint_value(value, bit_count=8),
+                            f"{field_name} should be a uint8 or NULL.")
+
+    def assert_valid_str(self, value, field_name):
+        """Asserts that the value is a non-empty string."""
+        asserts.assert_true(self.is_valid_str_value(value), f"{field_name} field should be a non-empty string")
+
     def get_test_steps(self, test: str) -> list[TestStep]:
         ''' Retrieves the test step list for the given test
